benjamins revised version...have fun :)
git-svn-id: file:///srv/local/data/cleanRepo/trunk@14270 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/fhuelsmann/src/main/java/playground/fhuelsmann/emission/analyseModul.java
@@ -25,9 +25,9 @@ import org.matsim.api.core.v01.Id;
 
 public interface analyseModul {
 	
-	public void calculateEmissionsPerPerson(final double travelTime, final Id personId, final double averageSpeed, final int roadType, final int freeVelocity, final double distance,HbefaObject[][] hbefaTable);
+	public void calculateEmissionsPerPerson(final double travelTime, final Id personId, final double averageSpeed, final int roadType, final double freeVelocity, final double distance,HbefaObject[][] hbefaTable);
 	
-	public void calculateEmissionsPerLink(final double travelTime, final Id linkId, final double averageSpeed, final int roadType, final int freeVelocity, final double distance,HbefaObject[][] hbefaTable);
+	public void calculateEmissionsPerLink(final double travelTime, final Id linkId, final double averageSpeed, final int roadType, final double freeVelocity, final double distance,HbefaObject[][] hbefaTable);
 	
 //	public void calculateColdEmissionsPerLink(final double activityTime, final Id personId,HbefaObject[][] hbefaTable);
 }
